fixsize

+++ python/ray/tune/examples/pbt_dcgan_mnist/common.py
@@ -185,13 +185,14 @@ def train(netD, netG, optimG, optimD, criterion, dataloader, iteration, device,
         netD.zero_grad()
         real_cpu = data[0].to(device)
         b_size = real_cpu.size(0)
-        label = torch.full((b_size, ), real_label, device=device)
+        label = torch.full(
+            (b_size, ), real_label, dtype=torch.float, device=device)
         output = netD(real_cpu).view(-1)
         errD_real = criterion(output, label)
         errD_real.backward()
         D_x = output.mean().item()
 
-        noise = torch.randn(b_size, nz, 1, 1, device=device)
+        noise = torch.randn(b_size, nz, 1, 1, dtype=torch.float, device=device)
         fake = netG(noise)
         label.fill_(fake_label)
         output = netD(fake.detach()).view(-1)
